Call the learning content replication directly from the main script
